Clarify infobox labels.
Change the "First Game" label to "Scorigami Achieved" to be clearer.
Every subsequent game with the same score is not a scorigami and
what's most notable about the first game with this score is that it
has created or achieved a scorigami. Change the "Last Game" label to
"Latest Game" since, for many of these scores, this will not be the
last game to have this score, it's only the latest game.